+++ tests/ResetPasswordTestKernel.php
@@ -91,7 +91,7 @@ class ResetPasswordTestKernel extends Kernel
                     'mappings' => [
                         'App' => [
                             'is_bundle' => false,
-                            'type' => self::shouldUseAttributes() ? 'attribute' : 'annotation',
+                            'type' => 'attribute',
                             'dir' => 'tests/Fixtures/Entity/',
                             'prefix' => 'SymfonyCasts\Bundle\ResetPassword\Tests\Fixtures\Entity',
                             'alias' => 'App',
@@ -140,10 +140,5 @@ class ResetPasswordTestKernel extends Kernel
     public function getLogDir(): string
     {
         return sys_get_temp_dir().'/logs'.spl_object_hash($this);
-    }
-
-    public static function shouldUseAttributes(): bool
-    {
-        return Kernel::VERSION_ID >= 70000;
     }
 }